requirements.txt
Flask==2.0.2
PyJWT==2.3.0
Werkzeug==2.0.2
requests==2.26.0
sentry-sdk>=1.4.3